Rename the package from `io.spine.test` to `io.spine.testing`.
