+++ base/src/test/java/io/spine/base/ThrowableMessageShould.java
@@ -22,7 +22,7 @@ package io.spine.base;
 import com.google.protobuf.Any;
 import com.google.protobuf.GeneratedMessageV3;
 import com.google.protobuf.util.Timestamps;
-import io.spine.test.Tests;
+import io.spine.testing.Tests;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
@@ -30,7 +30,7 @@ import org.junit.rules.ExpectedException;
 
 import java.util.Optional;
 
-import static io.spine.test.TestValues.newUuidValue;
+import static io.spine.testing.TestValues.newUuidValue;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertSame;
