+++ base/src/test/java/io/spine/io/PropertyFilesShould.java
@@ -22,7 +22,7 @@ package io.spine.io;
 
 import org.junit.Test;
 
-import static io.spine.test.Tests.assertHasPrivateParameterlessCtor;
+import static io.spine.testing.Tests.assertHasPrivateParameterlessCtor;
 
 /**
  * @author Alexander Yevsyukov
