+++ base/src/test/java/io/spine/json/JsonShould.java
@@ -26,7 +26,7 @@ import com.google.protobuf.StringValue;
 import com.google.protobuf.util.JsonFormat;
 import io.spine.json.given.Node;
 import io.spine.json.given.WrappedString;
-import io.spine.test.Tests;
+import io.spine.testing.Tests;
 import io.spine.type.KnownTypes;
 import io.spine.type.TypeUrl;
 import org.junit.Rule;
@@ -40,7 +40,7 @@ import static io.spine.json.Json.fromJson;
 import static io.spine.json.Json.toCompactJson;
 import static io.spine.json.Json.toJson;
 import static io.spine.protobuf.TypeConverter.toMessage;
-import static io.spine.test.Tests.assertHasPrivateParameterlessCtor;
+import static io.spine.testing.Tests.assertHasPrivateParameterlessCtor;
 import static java.lang.String.format;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
