+++ base/src/test/java/io/spine/logging/LoggingShould.java
@@ -29,7 +29,7 @@ import org.slf4j.Logger;
 
 import java.util.function.Supplier;
 
-import static io.spine.test.Tests.assertHasPrivateParameterlessCtor;
+import static io.spine.testing.Tests.assertHasPrivateParameterlessCtor;
 import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNotSame;
