+++ base/src/test/java/io/spine/money/MoneyUtilShould.java
@@ -22,7 +22,7 @@ package io.spine.money;
 
 import org.junit.Test;
 
-import static io.spine.test.Tests.assertHasPrivateParameterlessCtor;
+import static io.spine.testing.Tests.assertHasPrivateParameterlessCtor;
 import static org.junit.Assert.assertEquals;
 
 public class MoneyUtilShould {
