+++ base/src/test/java/io/spine/net/InternetDomainsShould.java
@@ -25,7 +25,7 @@ import org.junit.Test;
 
 import static io.spine.net.InternetDomains.pattern;
 import static io.spine.net.InternetDomains.valueOf;
-import static io.spine.test.Tests.assertHasPrivateParameterlessCtor;
+import static io.spine.testing.Tests.assertHasPrivateParameterlessCtor;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
