+++ base/src/test/java/io/spine/net/PatternsShould.java
@@ -22,7 +22,7 @@ package io.spine.net;
 
 import org.junit.Test;
 
-import static io.spine.test.Tests.assertHasPrivateParameterlessCtor;
+import static io.spine.testing.Tests.assertHasPrivateParameterlessCtor;
 
 /**
  * @author Alexander Yevsyukov
