+++ base/src/test/java/io/spine/net/SchemasShould.java
@@ -23,7 +23,7 @@ package io.spine.net;
 import com.google.common.testing.NullPointerTester;
 import org.junit.Test;
 
-import static io.spine.test.Tests.assertHasPrivateParameterlessCtor;
+import static io.spine.testing.Tests.assertHasPrivateParameterlessCtor;
 import static org.junit.Assert.assertEquals;
 
 /**
