+++ base/src/test/java/io/spine/protobuf/MessageFieldExceptionShould.java
@@ -24,10 +24,10 @@ import com.google.protobuf.Empty;
 import com.google.protobuf.StringValue;
 import com.google.protobuf.Timestamp;
 import io.spine.base.Time;
-import io.spine.test.Tests;
+import io.spine.testing.Tests;
 import org.junit.Test;
 
-import static io.spine.test.TestValues.newUuidValue;
+import static io.spine.testing.TestValues.newUuidValue;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
