+++ base/src/test/java/io/spine/type/CommandMessageShould.java
@@ -26,7 +26,7 @@ import io.spine.base.CommandMessage;
 import io.spine.base.given.CommandFromCommands;
 import org.junit.Test;
 
-import static io.spine.test.Tests.assertHasPrivateParameterlessCtor;
+import static io.spine.testing.Tests.assertHasPrivateParameterlessCtor;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
