+++ base/src/test/java/io/spine/util/ExceptionsShould.java
@@ -21,12 +21,12 @@
 package io.spine.util;
 
 import com.google.common.testing.NullPointerTester;
-import io.spine.test.TestValues;
+import io.spine.testing.TestValues;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
-import static io.spine.test.Tests.assertHasPrivateParameterlessCtor;
+import static io.spine.testing.Tests.assertHasPrivateParameterlessCtor;
 import static io.spine.util.Exceptions.newIllegalArgumentException;
 import static io.spine.util.Exceptions.newIllegalStateException;
 import static io.spine.util.Exceptions.unsupported;
