+++ testlib/src/test/java/io/spine/testing/VerifyShould.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.test;
+package io.spine.testing;
 
 import com.google.common.collect.ArrayListMultimap;
 import com.google.common.collect.FluentIterable;
@@ -40,39 +40,39 @@ import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.Callable;
 
-import static io.spine.test.Tests.assertHasPrivateParameterlessCtor;
-import static io.spine.test.Verify.assertBefore;
-import static io.spine.test.Verify.assertClassNonInstantiable;
-import static io.spine.test.Verify.assertContains;
-import static io.spine.test.Verify.assertContainsAll;
-import static io.spine.test.Verify.assertContainsEntry;
-import static io.spine.test.Verify.assertContainsKey;
-import static io.spine.test.Verify.assertContainsKeyValue;
-import static io.spine.test.Verify.assertEmpty;
-import static io.spine.test.Verify.assertEndsWith;
-import static io.spine.test.Verify.assertEqualsAndHashCode;
-import static io.spine.test.Verify.assertError;
-import static io.spine.test.Verify.assertInstanceOf;
-import static io.spine.test.Verify.assertItemAtIndex;
-import static io.spine.test.Verify.assertIterableEmpty;
-import static io.spine.test.Verify.assertIterableNotEmpty;
-import static io.spine.test.Verify.assertMapsEqual;
-import static io.spine.test.Verify.assertNegative;
-import static io.spine.test.Verify.assertNotContains;
-import static io.spine.test.Verify.assertNotContainsKey;
-import static io.spine.test.Verify.assertNotEmpty;
-import static io.spine.test.Verify.assertNotEquals;
-import static io.spine.test.Verify.assertNotInstanceOf;
-import static io.spine.test.Verify.assertPositive;
-import static io.spine.test.Verify.assertSetsEqual;
-import static io.spine.test.Verify.assertShallowClone;
-import static io.spine.test.Verify.assertSize;
-import static io.spine.test.Verify.assertStartsWith;
-import static io.spine.test.Verify.assertThrows;
-import static io.spine.test.Verify.assertThrowsWithCause;
-import static io.spine.test.Verify.assertZero;
-import static io.spine.test.Verify.denyContainsKey;
-import static io.spine.test.Verify.mangledException;
+import static io.spine.testing.Tests.assertHasPrivateParameterlessCtor;
+import static io.spine.testing.Verify.assertBefore;
+import static io.spine.testing.Verify.assertClassNonInstantiable;
+import static io.spine.testing.Verify.assertContains;
+import static io.spine.testing.Verify.assertContainsAll;
+import static io.spine.testing.Verify.assertContainsEntry;
+import static io.spine.testing.Verify.assertContainsKey;
+import static io.spine.testing.Verify.assertContainsKeyValue;
+import static io.spine.testing.Verify.assertEmpty;
+import static io.spine.testing.Verify.assertEndsWith;
+import static io.spine.testing.Verify.assertEqualsAndHashCode;
+import static io.spine.testing.Verify.assertError;
+import static io.spine.testing.Verify.assertInstanceOf;
+import static io.spine.testing.Verify.assertItemAtIndex;
+import static io.spine.testing.Verify.assertIterableEmpty;
+import static io.spine.testing.Verify.assertIterableNotEmpty;
+import static io.spine.testing.Verify.assertMapsEqual;
+import static io.spine.testing.Verify.assertNegative;
+import static io.spine.testing.Verify.assertNotContains;
+import static io.spine.testing.Verify.assertNotContainsKey;
+import static io.spine.testing.Verify.assertNotEmpty;
+import static io.spine.testing.Verify.assertNotEquals;
+import static io.spine.testing.Verify.assertNotInstanceOf;
+import static io.spine.testing.Verify.assertPositive;
+import static io.spine.testing.Verify.assertSetsEqual;
+import static io.spine.testing.Verify.assertShallowClone;
+import static io.spine.testing.Verify.assertSize;
+import static io.spine.testing.Verify.assertStartsWith;
+import static io.spine.testing.Verify.assertThrows;
+import static io.spine.testing.Verify.assertThrowsWithCause;
+import static io.spine.testing.Verify.assertZero;
+import static io.spine.testing.Verify.denyContainsKey;
+import static io.spine.testing.Verify.mangledException;
 import static java.util.Collections.singletonMap;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.fail;
