+++ tools/model-compiler/src/test/java/io/spine/tools/compiler/validation/ClassNamesShould.java
@@ -30,7 +30,7 @@ import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
 import static com.google.protobuf.DescriptorProtos.FieldDescriptorProto.getDefaultInstance;
-import static io.spine.test.Tests.assertHasPrivateParameterlessCtor;
+import static io.spine.testing.Tests.assertHasPrivateParameterlessCtor;
 import static io.spine.tools.compiler.validation.ClassNames.getClassName;
 import static io.spine.tools.compiler.validation.ClassNames.getStringClassName;
 import static io.spine.tools.compiler.validation.ClassNames.getValidatorMessageClassName;
