+++ tools/model-compiler/src/test/java/io/spine/tools/compiler/validation/MethodConstructorsShould.java
@@ -24,7 +24,7 @@ import com.google.common.testing.NullPointerTester;
 import com.squareup.javapoet.ClassName;
 import org.junit.Test;
 
-import static io.spine.test.Tests.assertHasPrivateParameterlessCtor;
+import static io.spine.testing.Tests.assertHasPrivateParameterlessCtor;
 import static io.spine.tools.compiler.validation.MethodConstructors.createConvertSingularValue;
 import static io.spine.tools.compiler.validation.MethodConstructors.createDescriptorStatement;
 import static io.spine.tools.compiler.validation.MethodConstructors.createValidateStatement;
