+++ tools/plugin-base/src/test/java/io/spine/tools/gradle/ProjectHierarchyShould.java
@@ -28,7 +28,7 @@ import org.junit.Test;
 import java.util.Set;
 
 import static com.google.common.collect.Sets.newHashSet;
-import static io.spine.test.Tests.assertHasPrivateParameterlessCtor;
+import static io.spine.testing.Tests.assertHasPrivateParameterlessCtor;
 import static java.util.Arrays.asList;
 import static java.util.Collections.emptySet;
 import static org.junit.Assert.assertEquals;
